+++ packages/stream_chat_flutter/lib/src/components/message_composer/message_composer_leading.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_core_flutter/stream_core_flutter.dart';
 
 /// A widget that shows the leading of the message composer.
 /// Uses the factory to show custom components or the default implementation.