+++ packages/stream_chat_flutter/lib/src/misc/thread_header.dart
@@ -62,7 +62,7 @@ class StreamThreadHeader extends StatelessWidget implements PreferredSizeWidget 
   final StreamAppBarStyle? style;
 
   @override
-  Size get preferredSize => const Size.fromHeight(kStreamHeaderHeight);
+  Size get preferredSize => const Size.fromHeight(kStreamToolbarHeight);
 
   @override
   Widget build(BuildContext context) {