+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -234,6 +234,9 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
   @override
   String get photosLabel => 'Photos';
 
+  @override
+  String get photosAndVideosLabel => 'Photos & Videos';
+
   String _getDay(DateTime dateTime) {
     final now = DateTime.now();
     final today = DateTime(now.year, now.month, now.day);