+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -235,6 +235,9 @@ class StreamChatLocalizationsEs extends GlobalStreamChatLocalizations {
   @override
   String get photosLabel => 'Fotos';
 
+  @override
+  String get photosAndVideosLabel => 'Fotos y vídeos';
+
   String _getDay(DateTime dateTime) {
     final now = DateTime.now();
     final today = DateTime(now.year, now.month, now.day);